requirements.txt
torch==2.2.0
pandas==2.2.0
numpy==1.26.4
matplotlib==3.8.2
scipy==1.12.0
PyYAML==6.0.1
tqdm==4.66.2
tables==3.9.2
jupyterlab==4.1.0
ipykernel==6.29.2
ipywidgets==8.1.2
scikit-learn==1.4.0
joblib==1.3.2